CHROMIUM: ASoC: rockchip: Add default MCLK and PLL for da7219 init
The suggestion that is from Adam Thomson who is the Dialog driver
developer for da7219:
--------
In the init() function where you make a call to da7219_aad_jack_det(),
prior to this I would suggest adding calls to snd_soc_dai_set_sysclk()
and snd_soc_dai_set_pll() to provide initial default MCLK settings to
the Codec. This means on start-up, and prior to any audio playback, you
have reasonable default MCLK and PLL settings for the Codec so that the
accessory detection works as expected, in particular for the HP detect
functionality.
--------

BUG=chrome-os-partner:54440
TEST=Can verify 3-pole jack

Change-Id: I6dd4ba035d31b7a098473f15190baeb9af2364fa
Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/365268
Commit-Ready: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>