CHROMIUM: iommu: Defer adding devices to groups
To make sure that the IOMMU drivers themselves are already probed, delay
adding devices to groups until they are attempted to be probed. This is
still a bit flaky, because the IOMMU driver must be registered before
other drivers (e.g. by using an earlier initcall), but at least can
handle cases when IOMMU needs another device to work, e.g. a power domain
as in Rockchip case.

BUG=chrome-os-partner:55068
TEST=Add logs, see that VOP and VPU are added to group and attached to domain properly

Change-Id: I48a07941e08b0ec809fdd3d0075b0d6b6877b658
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/366145
Reviewed-by: Owen Lin <owenlin@chromium.org>