CHROMIUM: arm64: dts: rockchip: Add supply to backlight
This patch adds the pp3300_disp supply to the backlight devicetree node.

Without this, the backlight uses a dummy regulator which is always on.
This, in turn, will cause the backlight power state to be initially set
to on. We don't want this behavior since we need to toggle the panel
regulator in order to read hpd/edid, and we don't want the backlight to
flash on/off everytime we do this.

With this patch, the backlight's initial state will be off and we'll
rely on the panel driver to enable/disable the backlight as appropriate.

BUG=None
TEST=Boot kevin with gerrit change 362183 and ensure no backlight flashes

Change-Id: I85ec510b6a37bf20638e1e2dfcb47e8eb06f2b10
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/366370
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Yakir Yang <ykk@rock-chips.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>