CHROMIUM: arm64: dts: rockchip: Add rk3399-gru-kevin-r6.dts
We've changed the dvfs range for kevin-r6 and we expect that the next
gru (gru-r2) will have the same range.  Therefor make that range the
default for rk3399-gru and adjust all old revs for the old range (with
the assumption that support for most old revs will eventually be
dropped).

BUG=chrome-os-partner:54888
TEST=dtc

Change-Id: I2a5bdd0e98d029bcd45ac9537df2393709da1f86
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/368349
Reviewed-by: David Schneider <dnschneid@chromium.org>