CHROMIUM: arm64: dts: rockchip: Fix pp3300_wifi_bt on rk3399-gru
The enable for pp3300_wifi_bt and pp1800_pcie is the same, but we had
different values.  I believe that this changed in an earlier rev and we
only updated one place.  Oops.

Fix this so that we don't end up driving gpio0_b0 (WIFI_HOST_WAKE_L)
high all the time.  I guess nobody noticed this?

BUG=chrome-os-partner:54026
TEST=Boot; WiFi still works.
TEST=Check /sys/kernel/debug/gpio and see no more gpio-8 high

Change-Id: I247991a2a96bd62fcbdc0bbfc44879bc0ed8e94d
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/368622
Reviewed-by: Brian Norris <briannorris@chromium.org>