CHROMIUM: HACK: clk: rockchip: don't use frac clocks for VOP on rk3399
In order to get low jitter on the pixel clock, testing indicates that
the fractional divider that can be a source to the VOPs needs to be _at
least_ "divide by 20".  Even with a source PLL of 2.4 GHz that gives us
a maximum pixel clock of 120 MHz.  This is not enough to support
all pixel clocks that we expect to need.

In theory we could be dynamic and use the fractional clock to make some
of the stranger clock frequencies that are < 120 MHz.  That would be
ideal, but at the moment it's pretty hard because:

1. CLK_SET_RATE_PARENT doesn't work on fractional clocks.
See http://crosreview.com/371641

2. We'd need to integrate the concept of jitter into the Rockchip clock
framework and VOP client (maybe using clk_get_accuracy?) so we could set
jitter requirements for the pixel clock and make sure that the fraction
clock isn't configured in a way to give too much jitter.

The above is a very difficult task.  Until all that is done, let's
simply hack out the connection to the fractional clock which will force
common clock framework to just never use it.

Change-Id: I5d11ae30a1bdc7f7fb580726249e7944689ad759
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/370779
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>