UPSTREAM: soc: rockchip: add header for ddr rate SIP interface
Add a header for the SIP interface defined to access the dcf controller
handling ddr rate changes on rk3399 (and most likely later socs).

This interface is shared between the clock driver as well as the
devfreq driver.

The SIP interface counterpart was merged from pull-request #684 [0]
into the upstream arm-trusted-firmware codebase.

[0] https://github.com/ARM-software/arm-trusted-firmware/pull/684

BUG=chrome-os-partner:54651
TEST=Boot from kevin

Change-Id: Iaffcbb588f605903d8686712ddabc8bda3341b29
Signed-off-by: Lin Huang <hl@rock-chips.com>
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Reviewed-on: https://chromium-review.googlesource.com/380318
Reviewed-by: Derek Basehore <dbasehore@chromium.org>