CHROMIUM: arm64: dts: assign a clock rate to VOP dclk
The VOP dclk should be enabled before writing registers, otherwise the
VOP will be blocked, all of the registers setting will not take effect.
This dclk is enabld in default value of register, and the core-boot or
u-boot may also enable it. But that can not guarantee dclk always be
enabled before vop_initialil. In RK3399, the dclk's parent is setted to
VPLL by default, but this VPLL probably be turned off before
vop_initialil, that cause nothing output from VOP. Assign a rate to the
vop dclk, can make sure this dclk get a valid parent and a good clock
rate to make all of the registers writing take effect.

BUG=chrome-os-partner:53565
TEST=eDP works without Type-C DP connection, when the cdn-dp is enabled
in dts.

Change-Id: Ic53c762032e76714f14edb7ea10492648053b331
Signed-off-by: Yakir Yang <ykk@rock-chips.com>
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/374262
Reviewed-by: Derek Basehore <dbasehore@chromium.org>