CHROMIUM: drm/atomic_helper: Do not duplicate fences in plane state
If a fence has been added to plane state for one commit, it is only
valid for that commit and should not be used for subsequent commit.

It might seem harmless, but if a commit is created before preceding
commit completes waiting for its fences, then we end up with the new
commit getting those fences assigned, even if it does not have to wait
for them (fb is different or plane is being disabled).

This patch fixes the problem by NULLing out the fence pointer, which
is safe, because if the new commit has to wait for some fences, they
will be assigned later in drm_atomic_add_implicit_fences().

BUG=chrome-os-partner:57092
TEST=while true; do backlight_dbus_tool --set --percent=0 && sleep 8 && backlight_dbus_tool --set --percent=100 && sleep 3 ; done

Change-Id: I74a9ad4c20adfb7db02ce11d5ab0aa7af8eff899
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/379691
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>