FROMLIST: drm/bridge: analogix_dp: Improve panel on time
In order to reduce the time required to turn on the panel, this patch
makes 2 assumptions:
1- In detect(): if there's a panel, we're connected.
2- In get_modes(): if there's a panel, let the panel driver decide if
   it should prepare/unprepare in order to get the modes.

The first is straightforward, and shouldn't need further explanation. The
second should eliminate the prepare/unprepare delays from get_modes() in
most cases, since panels generally hardcode their modes in the driver as
opposed to reading EDID. If a panel does need to read EDID, it should be
responsible for ensuring it's in a state in which it can.

BUG=chrome-os-partner:57063
TEST=Tested on kevin, panel on time goes from ~800ms to ~250ms

[from https://patchwork.freedesktop.org/patch/109162/]
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Ic6283608a6f03c160ec184cfdb59853db3599c67
Reviewed-on: https://chromium-review.googlesource.com/381872
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>