FROMLIST: PCI: rockchip: Increase the Max Credit update interval.
This increases the likelihood of link state to automatically go to L1
and save some power.

The default credit update interval of 7.5 us results in the rootport
sending UpdateFC-P and UpdateFC-NP packets too often, thus resulting
in the link never going to L1, and always staying in L0/L0s. The
value 24 us was chosen after some experiments and peeking over the
PCIe bus to see that we do enter L1 substate when there is not enough
traffic on the PCIe bus.

Signed-off-by: Rajat Jain <rajatja@google.com>
Acked-by: Shawn Lin <shawn.lin@rock-chips.com>
(am from https://patchwork.kernel.org/patch/9347261/)

BUG=chrome-os-partner:56848
TEST=Check wifi power on rk3399 kevin (i.e., entering L1 link PM state);
     Check PCIE_CLIENT_POWER_STATUS (`mem r 0xfd000008`) to see we're in
     L1 when link is relatively idle

Change-Id: I848f73e638af2e4f6999e2c598b1132662e6a1bb
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/386441