FROMLIST: mwifiex: fix command timeout problem seen in stress tests
It is observed that if single tid 6 packet comes among with massive tid 0
packets, tid 6 packet may stay in it's queue and will never be
transmited. This is because wmm.highest_queued_prio will be set to 2
during transmission of tid 0 packets As a result, main work thread
keeps on looping without serving that packet. In this case, if command
has downloaded to firmware, driver doesn't process it's response causing
command timeout.

This patch will reset highest_queued_prio if packets exist in data
queue, and try to find a ra_list for current private.

Signed-off-by: Xinming Hu <huxm@marvell.com>
Signed-off-by: Amitkumar Karwar <akarwar@marvell.com>
(am from https://patchwork.kernel.org/patch/9353821/)

BUG=chrome-os-partner:58164, chrome-os-partner:52627
TEST=repeated, simultaneous iperf and `iw mlan0 scan` tests

Change-Id: I07daea91ac8c076f8435c46d9eff55f7dde1a2da
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/393089
Reviewed-by: Abhishek Bhardwaj <abhishekbh@google.com>