CHROMIUM: drm/rockchip: Bypass atomic helpers for cursor updates
Moving the driver to atomic helpers regressed cursor responsiveness,
because cursor updates need their own atomic commits, which have to be
serialized with other commits, that might include fence waits. To avoid
this, in certain conditions, we can bypass the atomic helpers for legacy
cursor update IOCTLs. Currently the conditions are:
 - no asynchronous mode setting commit pending,
 - no asynchronous commit that updates the cursor plane is pending.
With the above two conditions met, we know that the manual cursor state
update will not conflict with any scheduled update.

BUG=chrome-os-partner:56580
TEST=Trigger a lot of screen updates and play with cursor.

Change-Id: Ie1f883a529366d31e184c0a5304bf9063f5e5a59
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/394492
Reviewed-by: Sean Paul <seanpaul@chromium.org>