CHROMIUM: [media] rockchip-vpu: fix h264 register programming on rk3399-vdec
Do not need dpb reorder on rk3399 when h264 decoding.
and if reordering apply on rk3399, h264 decoding would be
corrupted.

And when videos in frame mode decoding, both top field in used
and bottom field in used flag need to be set in all the
reference frames.

BUG=chrome-os-partner:58065
TEST=vimeo.com/122800995
	vimeo.com/178644541

Change-Id: I36d5710359e42d0809557d59a5cfdcc96218cda5
Signed-off-by: Alpha Lin <alpha.lin@rock-chips.com>
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/392727
Commit-Ready: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>