FROMLIST: PCI: rockchip: correct the use of FTS mask
We're trying to mask out bits[23:8] while retaining [32:24, 7:0], but
we're doing the inverse. That doesn't have too much effect, since we're
setting all the [23:8] bits to 1, and the other bits are only relevant
for modes we're currently not using. But we should get this right.

Fixes: ca1989084054 ("PCI: rockchip: Fix wrong transmitted FTS count")
Signed-off-by: Brian Norris <briannorris@chromium.org>
(am from https://patchwork.kernel.org/patch/9383297/)

BUG=chrome-os-partner:57856, chrome-os-partner:56693
TEST=kevin wifi

Change-Id: I0fefd1d6b1e799168c32b338b6feabb9a2f9f2bc
Reviewed-on: https://chromium-review.googlesource.com/399910
Commit-Ready: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>