platform/x86: intel_cht_int33fe: Set supplied-from property on max17047 dev
On devices with the intel_cht_int33fe ACPI device use a max17047
fuel-gauge combined with a bq24272i charger, in order for the
fuel-gauge driver to correctly display charging / discharging status
it needs to know which charger is supplying the battery.

This commit sets the supplied-from device property to the name of the
bq24272i charger for this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>